prepare for chart in result

+++ application/views/templates/trans/index.php
@@ -391,37 +391,47 @@
           myDropzone.processQueue(); // Tell Dropzone to process all queued files.
         });
         this.on("addedfile", function(file){
-          <?php if ($this->session->userdata('reg') == 1) { ?>
-
-          if (file.name == "PBS 1.csv" ) {
-            //do nothing?
-          }
-          else {
-            alert('file '+file.name+' tidak diperkenankan untuk di upload.');
-            this.removeFile(file);
-          }
+          <?php if ($this->session->userdata('lvl') == 1) { ?>
+            if (file.name == "PBS 1.csv" || file.name == "PBS 3.csv" || file.name == "PBS 2.csv" ) {
+              //do nothing?
+            }
+            else {
+              alert('file '+file.name+' tidak diperkenankan untuk di upload.');
+              this.removeFile(file);
+            }
           <?php } ?>
+          <?php if ($this->session->userdata('lvl') == 2) { ?>
+            <?php if ($this->session->userdata('reg') == 1) { ?>
+            if (file.name == "PBS 1.csv" ) {
+              //do nothing?
+            }
+            else {
+              alert('file '+file.name+' tidak diperkenankan untuk di upload.');
+              this.removeFile(file);
+            }
+            <?php } ?>
 
-          <?php if ($this->session->userdata('reg') == 2) { ?>
+            <?php if ($this->session->userdata('reg') == 2) { ?>
 
-          if (file.name == "PBS 2.csv" ) {
-            //do nothing?
-          }
-          else {
-            alert('file '+file.name+' tidak diperkenankan untuk di upload.');
-            this.removeFile(file);
-          }
-          <?php } ?>
+            if (file.name == "PBS 2.csv" ) {
+              //do nothing?
+            }
+            else {
+              alert('file '+file.name+' tidak diperkenankan untuk di upload.');
+              this.removeFile(file);
+            }
+            <?php } ?>
 
-          <?php if ($this->session->userdata('reg') == 3) { ?>
+            <?php if ($this->session->userdata('reg') == 3) { ?>
 
-          if (file.name == "PBS 3.csv" ) {
-            //do nothing?
-          }
-          else {
-            alert('file '+file.name+' tidak diperkenankan untuk di upload.');
-            this.removeFile(file);
-          }
+            if (file.name == "PBS 3.csv" ) {
+              //do nothing?
+            }
+            else {
+              alert('file '+file.name+' tidak diperkenankan untuk di upload.');
+              this.removeFile(file);
+            }
+            <?php } ?>
           <?php } ?>
         });
         this.on("queuecomplete", function (progress, response) {
